# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iran Research Portal Observatory API",
    "version": "1.1.0",
    "description": "Read-only API for the ISC-aligned audit of 115 public research and technology portals and its 920 dimension-level evidence outcomes. RTPMI measures portal maturity, not scientific quality."
  },
  "servers": [{"url": "/api/v1"}],
  "paths": {
...
    "/evidence": {
      "get": {
        "summary": "Query the 920 dimension-level evidence outcomes",
        "parameters": [
          {"name": "university", "in": "query", "schema": {"type": "string"}},
          {"name": "dimension", "in": "query", "schema": {"type": "string", "enum": ["portalIdentity", "organization", "libraryDocuments", "laboratories", "industryTechnology", "informationTechnology", "systemsServices", "documentsRegulations"]}},
          {"name": "status", "in": "query", "schema": {"type": "string", "enum": ["verified", "observed-reference", "restricted", "unresolved"]}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 920}},
          {"name": "offset", "in": "query", "schema": {"type": "integer", "minimum": 0}}
        ],
        "responses": {"200": {"description": "Filtered evidence outcomes"}}
      }
    },
...
  }
}